backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-docx==1.1.0
pdfminer.six==20231228
python-pptx==0.6.23
Pillow==10.1.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-magic==0.4.27
python-slugify==8.0.1
python-json-logger==2.0.7
python-dateutil==2.8.2
email-validator==2.1.0.post1
python-dotenv==1.0.0
# oauth
authlib==1.3.1
# testing
pytest
pytest-asyncio
httpx>=0.27,<0.28
fastapi